several changes on modify and added EntityResponse on outputs

+++ src/main/java/controller/API.java
@@ -2,8 +2,10 @@ package controller;
 
 import lombok.*;
 import model.Product;
+import org.apache.coyote.Response;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import services.productServiceImp;
@@ -11,7 +13,7 @@ import services.productServiceImp;
 import java.time.LocalDateTime;
 import java.util.List;
 import model.Product;
-
+@CrossOrigin(origins = "*", maxAge = 3600)
 @RestController
 @RequestMapping("/inventory")
 public class API {
@@ -20,8 +22,8 @@ public class API {
     private productServiceImp productServiceImp;
 
     @GetMapping("/products/Hello")
-    public String home(){
-        return "Hola mundo";
+    public ResponseEntity<?> home(){
+        return ResponseEntity.ok().body("Hola mundo!");
     }
 
     @GetMapping("/products")
@@ -30,34 +32,59 @@ public class API {
     }
 
     @PostMapping("/products")
-    public void postProduct(@Validated @RequestBody Product product){
+    public ResponseEntity<?> postProduct(@Validated @RequestBody Product product){
         productServiceImp.addProducts(product);
+        return ResponseEntity.ok().body("The product has been added");
 
     }
 
     @PutMapping("/products/{id}")
-    public void updateProduct(@Validated @RequestParam Integer productId, @RequestParam String productName, @RequestParam String productCategory, @RequestParam Float productPrice, @RequestParam LocalDateTime productExpirationDate, @RequestParam Integer productQuantityStock){
-        productServiceImp.modifyProduct(productId,productName,productCategory,productPrice,productExpirationDate,productQuantityStock);
+    public ResponseEntity<?> updateProduct(@Validated @RequestParam Integer productId, @RequestBody Product uProduct ){
+        if(productServiceImp.searchProduct(productId) == null){
+            return ResponseEntity.notFound().build();
+        }
+        else{
+            productServiceImp.modifyProduct(productId,uProduct);
+            return ResponseEntity.ok().body("Product with id: "+productId+" has been modified.");
+        }
     }
 
 
     @PostMapping("/products/{id}/outofstock")
-    public void outOfStockProduct(@Validated @RequestParam Integer productId){
-        productServiceImp.outOfStockProduct(productId);
+    public ResponseEntity<?> outOfStockProduct(@Validated @RequestParam Integer productId){
+        if(productServiceImp.searchProduct(productId) == null){
+            return ResponseEntity.notFound().build();
+        }
+        else{
+            productServiceImp.outOfStockProduct(productId);
+            return ResponseEntity.ok().body("Product with id: "+productId+" has been marked as out of stock.");
+        }
     }
 
 
     @PutMapping("/products/{id}/instock")
-    public void inStockProduct(@Validated @RequestParam Integer productId, @RequestParam Integer productQuantityStock){
+    public ResponseEntity<?> inStockProduct(@Validated @RequestBody Integer productId, @RequestBody Integer productQuantityStock){
         if (productQuantityStock!=0){
-            productServiceImp.reStockProduct(productId,productQuantityStock);
+            if(productServiceImp.searchProduct(productId) == null){
+                return ResponseEntity.notFound().build();
+            }
+            else{
+                productServiceImp.reStockProduct(productId,productQuantityStock);
+                return ResponseEntity.ok().body("Product with id: "+productId+" has been restocked.");
+            }
         }
         else{
-
+            return ResponseEntity.badRequest().body("You can not restock 0 units on a product");
         }
     }
     @DeleteMapping("/products/{id}/delete")
-    public Boolean deleteProduct(@Validated @RequestParam Integer productId){
-        return productServiceImp.deleteProduct(productId);
+    public ResponseEntity<?> deleteProduct(@Validated @RequestParam Integer productId){
+        if(productServiceImp.searchProduct(productId) == null){
+            return ResponseEntity.notFound().build();
+        }
+        else{
+            productServiceImp.deleteProduct(productId);
+            return ResponseEntity.ok().body("Product with id: "+productId+" has been ceased to exist.");
+        }
     }
 }
